+++ xuance/torch/runners/runner_sc2.py
@@ -34,7 +34,7 @@ class SC2_Runner(Runner_Base):
         return win_rate, allies_dead_ratio, enemies_dead_ratio
 
     def test_episodes(self, test_T, n_test_runs):
-        test_scores = np.zeros(n_test_runs, np.float)
+        test_scores = np.zeros(n_test_runs, np.float32)
         last_battles_info = self.get_battles_info()
         for i_test in range(n_test_runs):
             running_scores = self.agents.run_episodes(None, n_episodes=self.n_envs, test_mode=True)
